Split clearing temp tables as an extra step, more speedup

+++ src/include/import/GTFSImporter.class.php
@@ -30,6 +30,7 @@ class GTFSImporter extends ZipImporter
 	public function startImport(string $datasetName, string $license, ?string $referenceDate, string $desc,
 	SplFileInfo $file, ?string $runUntil) : GTFSImporter
 	{
+		$this->log("Starte Import für neuen Datensatz $datasetName");
 		$this->addDataset(trim($_POST["name"]), $license, $referenceDate, $desc);
 		$this->setRunUntil($runUntil);
 		return $this->stepReadFiles($file);
@@ -40,6 +41,8 @@ class GTFSImporter extends ZipImporter
 		$this->setRunUntil($runUntil);
 		$importState = $this->getImportState();
 
+		$this->log("Nehme Import für Datensatz $datasetId wieder auf, Status ist $importState");
+
 		switch($importState)
 		{
 			case GTFSConstants::IMPORT_STATE_INIT:
@@ -49,6 +52,8 @@ class GTFSImporter extends ZipImporter
 			case GTFSConstants::IMPORT_STATE_FILTERED:
 				return $this->stepRefine();
 			case GTFSConstants::IMPORT_STATE_REFINED:
+				return $this->stepApply();
+			case GTFSConstants::IMPORT_STATE_APPLIED:
 				return $this->stepFinish();
 		}
 		$this->finish();
@@ -84,7 +89,6 @@ class GTFSImporter extends ZipImporter
 
 		return $this->stepFilter();
 	}
-
 	private function stepFilter() : GTFSImporter
 	{
 		if($this->shouldStop(GTFSConstants::IMPORT_STATE_FILES_READ))
@@ -114,7 +118,7 @@ class GTFSImporter extends ZipImporter
 
 		return $this->stepFinish();
 	}
-	private function stepFinish() : GTFSImporter
+	private function stepApply() : GTFSImporter
 	{
 		if($this->shouldStop(GTFSConstants::IMPORT_STATE_REFINED))
 		{
@@ -123,6 +127,18 @@ class GTFSImporter extends ZipImporter
 		}
 
 		$this->copyImportData();
+		$this->setImportState(GTFSConstants::IMPORT_STATE_APPLIED);
+		return $this;
+	}
+	private function stepFinish() : GTFSImporter
+	{
+		if($this->shouldStop(GTFSConstants::IMPORT_STATE_APPLIED))
+		{
+			$this->finish();
+			return $this;
+		}
+
+		$this->removeTempData();
 		$this->setImportState(GTFSConstants::IMPORT_STATE_COMPLETE);
 		$this->finish();
 		return $this;
@@ -312,7 +328,7 @@ class GTFSImporter extends ZipImporter
 		return $this;
 	}
 
-	public function removeUnusedData() : GTFSImporter
+	public function removeUnusedData(int $runCount = 0) : GTFSImporter
 	{
 		if(!$this->isRunning())
 		{
@@ -323,49 +339,75 @@ class GTFSImporter extends ZipImporter
 			$this->abort("Falsche Reihenfolge, kein Dataset angelegt.");
 		}
 
+		$totalCount = 0;
 		try
 		{
 			$this->log("Lösche ungültige routes...");
 			$count = $this->db->cleanupTable("routes", $this->datasetId, "routes cleanup", [["agency_id", "agency", "agency_id"]]);
 			$this->log("$count ungültige routes markiert.");
+			$totalCount += $count;
 
 			$this->log("Lösche ungültige trips...");
 			$count = $this->db->cleanupTable("trips", $this->datasetId, "trips cleanup", [["route_id", "routes", "route_id"]]);
 			$this->log("$count ungültige trips markiert.");
+			$totalCount += $count;
 
 			$this->log("Lösche ungültige stop_times...");
 			$count = $this->db->cleanupTable("stop_times", $this->datasetId, "stop_times cleanup", [["trip_id", "trips", "trip_id"], ["stop_id", "stops", "stop_id"]], true);
 			$this->log("$count gültige stop_times behalten.");
+			$totalCount += $count;
 
 			$this->log("Lösche unbenutzte stops...");
 			$count = $this->db->cleanupTable("stops", $this->datasetId, "stops cleanup reverse", [["stop_id", "stop_times", "stop_id"], ["stop_id", "stops", "parent_station"]]);
 			$this->log("$count unbenutzte stops markiert.");
+			$totalCount += $count;
 
 			$this->log("Lösche unbenutzte trips...");
 			$count = $this->db->cleanupTable("trips", $this->datasetId, "trips cleanup reverse", [["trip_id", "stop_times", "trip_id"]]);
 			$this->log("$count unbenutzte trips markiert.");
+			$totalCount += $count;
 
 			$this->log("Lösche unbenutzte routes...");
 			$count = $this->db->cleanupTable("routes", $this->datasetId, "routes cleanup reverse", [["route_id", "trips", "route_id"]]);
 			$this->log("$count unbenutzte routes markiert.");
+			$totalCount += $count;
 
 			$this->log("Lösche unbenutzte Kalenderdaten...");
 			$count = $this->db->cleanupTable("calendar", $this->datasetId, "calendar cleanup reverse", [["service_id", "trips", "service_id"]]);
 			$this->log("$count unbenutzte agencies markiert.");
+			$totalCount += $count;
 
 			$this->log("Lösche unbenutzte Kalenderausnahmen...");
 			$count = $this->db->cleanupTable("calendar_dates", $this->datasetId, "calendar_dates cleanup reverse", [["service_id", "trips", "service_id"]]);
 			$this->log("$count unbenutzte Kalenderausnahmen markiert.");
+			$totalCount += $count;
 
 			$this->log("Lösche unbenutzte agencies...");
 			$count = $this->db->cleanupTable("agency", $this->datasetId, "agencies cleanup reverse", [["agency_id", "routes", "agency_id"]]);
 			$this->log("$count unbenutzte agencies markiert.");
+			$totalCount += $count;
 		}
 		catch(DBException $e)
 		{
 			$this->abort("Datenbankfehler beim Löschen unbenutzter Daten.", $e);
 		}
 
+		if($totalCount <= 0)
+		{
+			$this->log("Keine Änderung nach Aufräumen, fertig.");
+			return $this;
+		}
+
+		$runCount++;
+		$maxCleanupRuns= 10;
+		if($runCount > $maxCleanupRuns)
+		{
+			$this->log("Aufräumen von Daten: Maximale Rekursionstiefe erreicht ($maxCleanupRuns)");
+		}
+		else
+		{
+			$this->removeUnusedData($runCount);
+		}
 		return $this;
 	}
 
@@ -409,10 +451,42 @@ class GTFSImporter extends ZipImporter
 					$count = $this->db->copyFromImportTable($tableName, $columns);
 					$percent = $importCount !== 0 ? $count / $importCount : 0;
 					$this->log("$count von $importCount Einträgen aus $tableName übernommen ($percent%).");
-
-					$this->log("Lösche temporäre Daten aus $importTableName...");
-					$this->db->deleteData($importTableName, $this->datasetId);
 				}
+			}
+			catch(DBException $e)
+			{
+				$this->abort("Datenbankfehler beim Übernehmen der Daten.", $e);
+			}
+		}
+
+		return $this;
+	}
+
+	protected function removeTempData() : GTFSImporter
+	{
+		if(!$this->isRunning())
+		{
+			$this->log("not running");
+			return $this;
+		}
+		if($this->datasetId === null)
+		{
+			$this->abort("Falsche Reihenfolge, kein Dataset angelegt.");
+		}
+
+		$this->log("files: ".json_encode($files));
+
+		foreach(GTFSFiles::FILES as $file)
+		{
+			$fileOptions = GTFSFiles::getFileOptions($file);
+
+			try
+			{
+				$tableName = $fileOptions->getTableName();
+				$importTableName = $this->db->getImportTableName($tableName);
+
+				$this->log("Lösche temporäre Daten aus $importTableName...");
+				$this->db->deleteData($importTableName, $this->datasetId);
 			}
 			catch(DBException $e)
 			{
